fix(claim-sync): reset to one localStorage
